refactor(fsd): cleanup
