+++ src/entities/game/api/schema.ts
@@ -1,15 +1,5 @@
 import { z } from "zod";
 
-export const SignInResponseSchema = z.object({
-	uid: z.string(),
-	name: z.string(),
-	email: z.email(),
-	profile_pic: z.url(),
-	new_account: z.boolean(),
-});
-
-export type SignInResponse = z.infer<typeof SignInResponseSchema>;
-
 export const CreateGameResponseSchema = z.object({
 	sessionId: z.string(),
 	uid: z.string(),
@@ -38,15 +28,6 @@ export const QuitGameResponseSchema = z.object({
 	success: z.boolean(),
 	error: z.string().optional(),
 });
-
-export const GetWalletResponseSchema = z.object({
-	coins: z.number().int(),
-	xp: z.number().int(),
-	success: z.boolean(),
-	error: z.string().optional(),
-});
-
-export type GetWalletResponse = z.infer<typeof GetWalletResponseSchema>;
 
 export const SkipMoveResponseSchema = z.object({
 	boards: z.array(z.number().int()),
