+++ src/entities/game/model/types.ts
@@ -1,7 +1,3 @@
-// This file contains all shared types and interfaces used across the application
-// Using TypeScript utility types to reduce duplication and improve maintainability
-
-import type { User } from "firebase/auth";
 import type {
 	MakeMoveResponse,
 	SkipMoveResponse,
@@ -55,44 +51,4 @@ export type UndoMoveResult =
 export interface QuitGameResponse {
 	success: boolean;
 	error?: string;
-}
-
-// ZUSTAND STORE TYPES
-
-export interface SoundStore {
-	bgMute: boolean;
-	bgVolume: number;
-	sfxMute: boolean;
-	sfxVolume: number;
-	setBgMute: (mute: boolean) => void;
-	setBgVolume: (vol: number) => void;
-	setSfxMute: (mute: boolean) => void;
-	setSfxVolume: (vol: number) => void;
-}
-export interface ProfileStore {
-	name: string;
-	email: string;
-	pic: string;
-	setName: (name: string) => void;
-	setEmail: (email: string) => void;
-	setPic: (pic: string) => void;
-}
-
-export interface UserStore {
-	user: User | null;
-	authReady: boolean;
-	setUser: (newUser: User | null) => void;
-	setAuthReady: (v: boolean) => void;
-}
-
-export interface CoinStore {
-	coins: number;
-	setCoins: (newCoins: number) => void;
-	optimisticAddCoins: (amount: number) => void;
-}
-
-export interface XPStore {
-	XP: number;
-	setXP: (newXP: number) => void;
-	optimisticAddXP: (amount: number) => void;
 }
