+++ src/features/authenticate-user/api/schema.ts
@@ -1,0 +1,11 @@
+import { z } from "zod";
+
+export const SignInResponseSchema = z.object({
+	uid: z.string(),
+	name: z.string(),
+	email: z.email(),
+	profile_pic: z.url(),
+	new_account: z.boolean(),
+});
+
+export type SignInResponse = z.infer<typeof SignInResponseSchema>;
