+++ src/features/manage-user-profile/model/profileStore.ts
@@ -1,0 +1,25 @@
+import { create } from "zustand";
+import { persist } from "zustand/middleware";
+
+interface ProfileStore {
+	name: string;
+	email: string;
+	pic: string;
+	setName: (name: string) => void;
+	setEmail: (email: string) => void;
+	setPic: (pic: string) => void;
+}
+
+export const useProfile = create<ProfileStore>()(
+	persist(
+		(set) => ({
+			name: "player",
+			email: "empty@empty.empty",
+			pic: "empty.empty",
+			setName: (name) => set({ name }),
+			setEmail: (email) => set({ email }),
+			setPic: (pic) => set({ pic }),
+		}),
+		{ name: "profile-settings" },
+	),
+);
