+++ src/features/manage-wallet/api/schema.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+
+export const GetWalletResponseSchema = z.object({
+	coins: z.number().int(),
+	xp: z.number().int(),
+	success: z.boolean(),
+	error: z.string().optional(),
+});
+
+export type GetWalletResponse = z.infer<typeof GetWalletResponseSchema>;
