+++ src/widgets/profile-modal/ui/ProfileModal.tsx
@@ -1,9 +1,6 @@
-import {
-	useCoins,
-	useProfile,
-	useUser,
-	useXP,
-} from "@/features/app-state/model/stores";
+import { useUser } from "@/features/authenticate-user/model/userStore";
+import { useProfile } from "@/features/manage-user-profile/model/profileStore";
+import { useCoins, useXP } from "@/features/manage-wallet/model/walletStore";
 import ModalOverlay from "@/widgets/modal-overlay/ui/ModalOverlay";
 import ProfileAvatar from "@/widgets/profile-avatar/ui/ProfileAvatar";
 import { ProfileButton } from "@/widgets/profile-button/ui/ProfileButton";
