+++ src/widgets/splash-title-block/ui/SplashTitleBlock.tsx
@@ -1,0 +1,12 @@
+export default function SplashTitleBlock() {
+	return (
+		<div className="text-center mb-8 animate-slide-up">
+			<div className="font-pixel text-primary text-2xl pixel-text-shadow mb-4">
+				NOTAKTO
+			</div>
+			<div className="font-pixel text-[8px] text-cream-dim">
+				NO TIES · ALWAYS A WINNER
+			</div>
+		</div>
+	);
+}
